feat(ui): optimize micro-interactions and fix draft badge display
- Why:
  - Improve tactile user experience through immediate visual feedback.
  - Reduce information noise by adhering to a minimalist design approach.

- What:
  - Add completion feedback (icon change) for batch copy operations with a 1000ms auto-close delay.
  - Implement a 400ms "resolve" animation (strikethrough and reduced opacity) before final processing.
  - Restrict "NEW" badge visibility to Note types in the right pane, hiding it for Drafts.
